requirements.txt
Pillow
requests
urllib3
tqdm
aiohttp
aiofiles

pytest
pytest-cov
pre-commit
